Create temp notebooks dir only locally

+++ tests/boilercv_tests/conftest.py
@@ -40,6 +40,8 @@ def _project_session_path(tmp_path_factory):
 @pytest.fixture(autouse=True, scope="session")
 def _test_nbs():
     """Set test notebook directory, cleaning up afterwards."""
+    if environ.get("CI"):
+        return
     TEST_TEMP_NBS.mkdir(exist_ok=True)
     yield
     rmtree(TEST_TEMP_NBS)
